eqe_mvc/requirements.txt
# Core GUI and Plotting
PySide6>=6.0.0
matplotlib>=3.5.0

# Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Device Communication
pyvisa>=1.11.0
pyserial>=3.5

# Optional: Additional development dependencies
# pytest>=6.0.0  # for testing
# black>=21.0.0  # for code formatting
# flake8>=3.9.0  # for linting